Rename contact_types package to items
The name contact_types was misleading since the classes in this package
are used to create different types of items, not contacts.
